Add CommandRPC support
	submitCommand for command submit.
	getCommandInfo for getting commandinfo with given commandID.
	listCommandinfs for getting all commands executing.
	activateCommand for activating(insert it to queue) an given command.
	deleteCommand for delete a command from queue, if this command is running, then kill it.

+++ hadoop-smart-project/src/main/java/org/apache/hadoop/smart/sql/CommandInfo.java
@@ -55,7 +55,7 @@ public class CommandInfo {
     return cid;
   }
 
-  public void setCid(int cid) {
+  public void setCid(long cid) {
     this.cid = cid;
   }
